# docs/swagger.yaml
basePath: /
definitions:
  requests.RefreshSession:
    properties:
      accessToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHBpcmVzQXQiOiIxMjM0NTY3ODkwIiwidXNlcklkIjoiMSIsImlzcyI6IlRPRE8ifQ.K-6Tzcaoae1Cj7jbIMalrtsLXZFrAlg_F_XLegWGo7o
        type: string
      refreshToken:
        example: $2a$10$9UKV92GI6504S7RpPPZApe1Llp3fyOS7TH4tQC9ty6OQLxcjIT8uC
        type: string
    required:
    - accessToken
    - refreshToken
    type: object
  requests.SignIn:
    properties:
      email:
        example: example@mail.ru
        type: string
      password:
        example: 123superPassword
        type: string
    required:
    - email
    - password
    type: object
  requests.SignUp:
    properties:
      email:
        example: example@mail.ru
        type: string
      password:
        example: 123superPassword
        type: string
    required:
    - email
    - password
    type: object
  responses.Session:
    properties:
      accessToken:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHBpcmVzQXQiOiIxMjM0NTY3ODkwIiwidXNlcklkIjoiMSIsImlzcyI6IlRPRE8ifQ.K-6Tzcaoae1Cj7jbIMalrtsLXZFrAlg_F_XLegWGo7o
        type: string
      expiresAt:
        example: 1592572800
        type: integer
      refreshToken:
        example: $2a$10$9UKV92GI6504S7RpPPZApe1Llp3fyOS7TH4tQC9ty6OQLxcjIT8uC
        type: string
    type: object
  responses.SignIn:
    properties:
      id:
        example: "2"
        type: string
      session:
        $ref: '#/definitions/responses.Session'
    type: object
  responses.SignUp:
    properties:
      id:
        example: "2"
        type: string
      session:
        $ref: '#/definitions/responses.Session'
    required:
    - id
    - session
    type: object
  responses.User:
    properties:
      email:
        type: string
      id:
        type: string
      registrationDate:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: service for auth users
  title: Auth Service
  version: 0.0.1
paths:
  /auth/session/logout:
    post:
      description: запрос на закрытие сессий пользователя по их id с использованием
        токена, переданного в заголовке "Authorization"
      parameters:
      - description: access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
        "401":
          description: некорректный access token
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: запрос на закрытие сессий пользователя
  /auth/signin:
    post:
      consumes:
      - application/json
      description: вход в аккаунт с использованием email + пароль для получения токенов
      parameters:
      - description: структура запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.SignIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SignIn'
        "400":
          description: некорректный формат запроса
          schema:
            type: string
        "401":
          description: неправильный пароль
          schema:
            type: string
        "404":
          description: пользователь не найден
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: вход в аккаунт
  /auth/token/{user_id}:
    get:
      consumes:
      - application/json
      description: создание токенов по id пользователя
      parameters:
      - description: path format
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Session'
        "400":
          description: некорректный формат запроса
          schema:
            type: string
        "404":
          description: пользователь не найден
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: создание токенов
  /auth/token/update:
    post:
      consumes:
      - application/json
      description: возвращает новую пару токенов при отправке старой пары и при условии
        их валидности
      parameters:
      - description: request format
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.RefreshSession'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Session'
        "400":
          description: некорректный формат запроса
          schema:
            type: string
        "401":
          description: невалидная пара токенов, либо истекший refresh token
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: обновление сессии
  /auth/user:
    get:
      description: запрос на получение пользователя с использованием токена, переданного
        в заголовке "Authorization"
      parameters:
      - description: access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.User'
        "401":
          description: некорректный access token
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: запрос на получение пользователя пользователя
  /signup:
    post:
      consumes:
      - application/json
      description: регистрация нового пользователя
      parameters:
      - description: структура запрос
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.SignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SignUp'
        "400":
          description: некорректный формат запроса
          schema:
            type: string
        "409":
          description: пользователь уже существует
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: регистрация нового пользователя
swagger: "2.0"